server/Python_Examples/agent_sdk_bobbys_table/requirements.txt
Flask
Flask_SQLAlchemy
Flask_HTTPAuth
python-dotenv
signalwire-agents
gunicorn
requests
stripe
pytz
